### 2025-02-26 6f6c47b
fix invalid DOIs in bib file
