+++ joss/paper.bib
@@ -13,7 +13,7 @@
     number = {11},
     pages = {e2023MS003686},
     keywords = {tropical cyclones, hazard, downscaling, open-source},
-    doi = {https://doi.org/10.1029/2023MS003686},
+    doi = {10.1029/2023MS003686},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2023MS003686},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2023MS003686},
     note = {e2023MS003686 2023MS003686},
@@ -158,7 +158,7 @@
     number = {1},
     pages = {165-186},
     keywords = {tropical cyclones, earth system modeling, TC-ocean interactions},
-    doi = {https://doi.org/10.1002/2017MS001199},
+    doi = {10.1002/2017MS001199},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/2017MS001199},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/2017MS001199},
     year = {2018}
@@ -227,7 +227,7 @@
     number = {12},
     pages = {e2020GL092244},
     keywords = {CMIP6, MJO, PCA, spectral, wavelet},
-    doi = {https://doi.org/10.1029/2020GL092244},
+    doi = {10.1029/2020GL092244},
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2020GL092244},
     eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2020GL092244},
     note = {e2020GL092244 2020GL092244},
@@ -256,7 +256,7 @@
     pages = {100269},
     year = {2020},
     issn = {2212-0947},
-    doi = {https://doi.org/10.1016/j.wace.2020.100269},
+    doi = {10.1016/j.wace.2020.100269},
     url = {https://www.sciencedirect.com/science/article/pii/S2212094719302439},
     author = {Yeon-Hee Kim and Seung-Ki Min and Xuebin Zhang and Jana Sillmann and Marit Sandstad},
     keywords = {Climate extremes, ETCCDI indices, CMIP6, Model evaluation, GEV analysis},
@@ -338,7 +338,7 @@
     number = {730},
     pages = {1999-2049},
     keywords = {climate reanalysis, Copernicus Climate Change Service, data assimilation, ERA5, historical observations},
-    doi = {https://doi.org/10.1002/qj.3803},
+    doi = {10.1002/qj.3803},
     url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.3803},
     eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.3803},
     year = {2020}
