### 2025-03-14 5f3b51a
fix: resolve indexing issue with short velocity time series
