### 2025-03-17 1fab67d
shorten the paper to meet the length required by JOSS
