### 2025-03-26 eeec087
continue reducing the length of the paper
