+++ joss/paper.md
@@ -53,7 +53,7 @@ To sample rare (most intense) rainfall events that are often of great societal i
 # Statement of need
 Tropical cyclones (TCs) -- that is, hurricanes and tropical storms -- are among the most destructive weather events, causing massive economic and human losses worldwide each year [@Krichene:2023]. In the United States, hurricanes can trigger a surge of deaths long after the storms through complex chains of lasting impacts [@Young:2024]. Much of the damage caused by TCs is done by water – particularly by torrential rainfall and subsequent flooding [@Shi:2024;@Zhang:2018]. Accurately capturing TC rainfall characteristics at high spatial (e.g., <10 km) and temporal (e.g, hourly) resolution is therefore of critical importance. Moreover, a growing body of evidence suggests that TC rainfall is becoming more intense under a warming climate due to the Clapeyron–Clausius scaling of water vapor in the atmosphere [@Held:2006], increasing the likelihood of extreme rainfall and flooding [@Zhu:2021]. Given the societal consequences of TCs, it is crucial to understand not only TC rainfall risk in the current climate, but also how the risk might evolve with warming. Advancing tools and models to accurately and efficiently quantify these risks is of great significance.
 
-The ability of GCMs to simulate climate extremes has been substantially improved over the past few decades, primarily those participating in the Coupled Model Intercomparison Project phase 6 or CMIP6 [@Eyring:2016]. These climate models have become one of the main tools used to explore the effect of global warming on precipitation and climate variability [@Emanuel:2021;@Le:2021;@Le:2023]. While high-resolution GCMs have improved the representation of TCs [@Haarsma:2016;@Li:2018], they remain too computationally expensive for risk analysis, which requires robust sampling of extreme rainfall events. `pyTCR` responds to this need for an easy to use and efficient tool that facilitates TC-driven rainfall analysis across scales. It takes the advantage of a synthetic downscaling approach that can generate large ensembles of synthetic TCs at the basin (ocean) scale based on comprehensive climate conditions from observations and reanalysis data and a large number of GCM simulations [@Emanuel:2008;@Lin:2023].
+The ability of GCMs to simulate climate extremes has been substantially improved over the past few decades. These climate models have become one of the main tools used to explore the effect of global warming on precipitation and climate variability [@Emanuel:2021;@Le:2021;@Le:2023]. While high-resolution GCMs have improved the representation of TCs [@Haarsma:2016;@Li:2018], they remain too computationally expensive for risk analysis, which requires robust sampling of extreme rainfall events. `pyTCR` responds to this need for an easy to use and efficient tool that facilitates TC-driven rainfall analysis across scales. It takes the advantage of a synthetic downscaling approach that can generate large ensembles of synthetic TCs at the basin (ocean) scale based on comprehensive climate conditions from observations and reanalysis data and a large number of GCM simulations [@Emanuel:2008;@Lin:2023].
 `pyTCR` provides a fast and highly efficient tool for risk analysis related to TC rainfall.
 
 
@@ -76,7 +76,7 @@ where $w_f$ represents the velocity induced by surface frictional convergence, $
 
 # Examples
 
-`pyTCR` generates and provides access to a large set of synthetic rainfall events based on TC tracks. To help users learn key concepts and functions of `pyTCR`, we provide six `Jupyter` notebooks designed for training purposes:
+To help users learn key concepts and functions of `pyTCR`, we provide six `Jupyter` notebooks designed for training purposes:
 
 - Downloading and preprocessing TCs data
 - Visualizing and analyzing TC tracks
[…]
