### 2025-03-27 #7849 @jatkinson1000
Hi @levuvietphong

I'm taking this over as editor from AEiC @kthyng:

Your submission looks almost ready for review, and I'll repeat here the pre-review checklist from @kthyng with a couple of my own comments:

As @kthyng notes, [the paper is currently too long](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain), so please look at reducing it.
My advice would be to remove some of the mathematical details to instead be included in the documentation.

Whilst users can run the notebooks, it would be good to see some formal testing in the code for verification purposes, particularly of the core functionalities in `tcr/`. This is [strongly encouraged by JOSS](https://joss.readthedocs.io/en/latest/review_criteria.html#tests) and the reviewers may raise this. Please let me know if you would like pointing to guidance on software testing, e.g. using [pytest](https://docs.pytest.org/en/stable/).

Finally, I see that you have conda install instructions. However, I would reiterate @kthyng's comments above encouraging you to facilitate other options, and to look at [packaging](https://packaging.python.org/en/latest/) the code for easy distribution.

If you can start looking into these points and get back to me we can then start looking for reviewers.
