### 2025-03-27 1742f22
feat: migrate installation to PyPI and add contribution guides
