### 2025-03-27 96813b0
feat: add sphinx read-the-docs for documentation
