### 2025-03-27 bf4fe90
feat: add config and requirements for read-the-docs
