+++ source/index.rst
@@ -1,0 +1,18 @@
+.. pyTCR documentation master file, created by
+   sphinx-quickstart on Thu Mar 27 15:30:52 2025.
+   You can adapt this file completely to your liking, but it should at least
+   contain the root `toctree` directive.
+
+pyTCR documentation
+===================
+
+Add your content using ``reStructuredText`` syntax. See the
+`reStructuredText <https://www.sphinx-doc.org/en/master/usage/restructuredtext/index.html>`_
+documentation for details.
+
+This is a sentence
+
+.. toctree::
+   :maxdepth: 2
+   :caption: Contents:
+
