### 2025-03-28 3e8e568
feat: add documentation using jupyter book
