+++ README.md
@@ -25,7 +25,7 @@
 - [Acknowledgments](#people_hugging-acknowledgments)
 - [Contacts](#mailbox-contacts)
 
-## :books: Overview
+## :book: Overview
 pyTCR is a physics-based model developed in python to estimate rainfall induced by tropical cyclones (TCs). It is largely based on the TCR model described by [Zhu *et al.*, 2013](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1002/2013GL058284) and [Lu *et al.*, 2018](https://journals.ametsoc.org/view/journals/atsc/75/7/jas-d-17-0264.1.xml). PyTCR simulates convective TC rainfall by correlating the precipitation rate with the total upward velocity within the TC vortex. It integrates seamlessly with outputs from [a tropical cyclone downscaling model](https://github.com/linjonathan/tropical_cyclone_risk) (see [Lin *et al.,* 2023](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2023MS003686)) to generate detailed spatial-temporal rainfall patterns that align with hurricane tracks.
 
 ## :wrench: Installation
@@ -35,13 +35,14 @@ pyTCR is a physics-based model developed in python to estimate rainfall induced 
     pip install pyTCR
     ```
 
-2. To get the latest updates from the github repository:
+2. To get the latest (unstable) development version from GitHub:
     ```sh
     git clone https://github.com/levuvietphong/pyTCR.git
     cd pyTCR
     pip install .
     ```
-    Cloning the repository also gives you access to the tutorial in the notebooks folder.
+    > [!NOTE]
+    > This installs the most recent updates, which may include new features and bug fixes that are not yet available in the stable release. Cloning the repository also gives you access to the tutorial in the notebooks folder.
 
 ## :arrow_forward: Getting Started
 This repository provides a collection of [jupyter notebooks](https://github.com/levuvietphong/pyTCR/tree/main/notebooks) designed to facilitate the use of PyTCR for generating and visualizing rainfall and wind speed induced by TCs. These notebooks cover various use cases and data sources, providing comprehensive guidance for users.
@@ -62,7 +63,7 @@ This repository provides a collection of [jupyter notebooks](https://github.com/
 > Downscaled tropical cyclone tracks for various CMIP6 models (including `historical` and `ssp585` experiments) are available for download [here](https://web.corral.tacc.utexas.edu/setxuifl/tropical_cyclones/downscaled_cmip6_tracks).
 
 ## :framed_picture: Gallery
-The following figures illustrate the inputs used in pyTCR and results obtained from PyTCR:
+The following figures illustrate the inputs used and results obtained from PyTCR:
 
 - **Tropical cyclone tracks:** The figure below shows the tracks of tropical cyclones downscaled from outputs from E3SM v1.0 model, Reanalysis ERA5, and best track observations.
 
@@ -72,16 +73,19 @@ The following figures illustrate the inputs used in pyTCR and results obtained f
 
 ![](images/cumulative_rain_3storms.png)
 
-- **Time Series of Precipitation:** The figure below depicts the time series of precipitation at a specific location (29°N, 95°W) within the SETx region.
+- **Time Series of Precipitation:** The figure below depicts the time series of precipitation at a specific location (29.907°N, 94.900°W) within the SETx region.
 
 ![](images/rainfall_timeseries.png)
 
+
+## :books: Documentation
+The official documentation is hosted on [Read the Docs](https://pytcr.readthedocs.io/en/latest/).
 
 ## :page_facing_up: License
 Distributed under the MIT License. See [LICENSE](LICENSE) for more information.
 
 ## :rocket: Contributions
[…]

[…]
