### 2025-03-30 69617c1
doc: update documentation website and notebooks
