+++ docs/api/modules.rst
@@ -1,0 +1,7 @@
+pyTCR
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   pyTCR
