### 2025-03-30 4d4ec33
doc: update readthedocs yaml file for sphinx
