### 2025-03-30 5cf7bb2
doc: fix bibliography and disable external execution
