+++ docs/_config.yml
@@ -16,9 +16,9 @@ latex:
   latex_documents:
     targetname: book.tex
 
-# Add a bibtex file so that we can create citations
-bibtex_bibfiles:
-  - references.bib
+# # Add a bibtex file so that we can create citations
+# bibtex_bibfiles:
+#   - references.bib
 
 # Information about where the book exists on the web
 repository:
@@ -31,6 +31,9 @@ repository:
 html:
   use_issues_button: true
   use_repository_button: true
+
+execute:
+  execute_notebooks: "off"
 
 sphinx:
   config:
@@ -65,9 +68,8 @@ sphinx:
           "floor": ["\\lfloor#1\\rfloor", 1]
           "bmat": ["\\left[\\begin{array}"]
           "emat": ["\\end{array}\\right]"]  
-  bibtex_reference_style: author_year
-  bibtex_bibfiles:
-    - references.bib
+    bibtex_reference_style: author_year
+    bibtex_bibfiles: ["references.bib"]
 
   extra_extensions:
     - sphinx.ext.intersphinx
