### 2025-03-30 645bf93
doc: fix bibliography and disable external execution
