+++ docs/api/index.rst
@@ -1,13 +1,0 @@
-.. pyTCR documentation master file, created by
-   sphinx-quickstart on Thu Mar 27 16:01:44 2025.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
-API
-===
-
-.. toctree::
-   :maxdepth: 2
-   :caption: Contents:
-
-   modules
