+++ .readthedocs.yaml
@@ -28,4 +28,3 @@ sphinx:
 
 formats:
   - pdf
-  - epub
