### 2025-03-31 f0e0474
doc: update mathematical background and references
