+++ docs/references.bib
@@ -13,10 +13,7 @@
     number = {11},
     pages = {e2023MS003686},
     keywords = {tropical cyclones, hazard, downscaling, open-source},
-    doi = {https://doi.org/10.1029/2023MS003686},
-    url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2023MS003686},
-    eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2023MS003686},
-    note = {e2023MS003686 2023MS003686},
+    doi = {https://doi.org/10.1029/2023MS003686},    
     year = {2023}
 }
 
@@ -31,7 +28,6 @@
     number = "7",
     doi = "10.1175/JAS-D-17-0264.1",
     pages=      "2337 - 2358",
-    url = "https://journals.ametsoc.org/view/journals/atsc/75/7/jas-d-17-0264.1.xml"
 }
 
 @article{Zhu:2013,
@@ -43,8 +39,6 @@
     pages = {6225-6230},
     keywords = {tropical cyclones (3372), precipitation (1854), risk (4328), synthetic simulation (4316)},
     doi = {10.1002/2013GL058284},
-    url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1002/2013GL058284},
-    eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1002/2013GL058284},
     year = {2013}
 }
 
@@ -284,8 +278,7 @@
     volume = "72",
     number = "9",
     doi = "10.1175/JAS-D-15-0014.1",
-    pages=      "3647 - 3662",
-    url = "https://journals.ametsoc.org/view/journals/atsc/72/9/jas-d-15-0014.1.xml"
+    pages = "3647 - 3662",
 }
 
 @article{Emanuel:2017,
@@ -297,8 +290,6 @@
     pages = {12681-12684},
     year = {2017},
     doi = {10.1073/pnas.1716222114},
-    URL = {https://www.pnas.org/doi/abs/10.1073/pnas.1716222114},
-    eprint = {https://www.pnas.org/doi/pdf/10.1073/pnas.1716222114},
 }
 
 @Article{Mendelsohn:2012,
@@ -339,8 +330,6 @@
     pages = {1999-2049},
     keywords = {climate reanalysis, Copernicus Climate Change Service, data assimilation, ERA5, historical observations},
     doi = {https://doi.org/10.1002/qj.3803},
-    url = {https://rmets.onlinelibrary.wiley.com/doi/abs/10.1002/qj.3803},
-    eprint = {https://rmets.onlinelibrary.wiley.com/doi/pdf/10.1002/qj.3803},
     year = {2020}
 }
 
