### 2025-04-02 3d47e5c
feat: add pytest for various tcr functions
