### 2025-04-02 e083d1a
doc: update docstring and api function of dirdata
