+++ docs/api/pyTCR.rst
@@ -1,5 +1,5 @@
-API 
-===
+API Reference
+=============
 
 Functions and Submodules
 ------------------------
