### 2025-04-02 58745ed
doc: revise the example part to shorten the paper
