### 2025-04-02 62b2d3e
doc: minor update on the introduction and version
