### 2025-04-03 #7849 @levuvietphong
Hi @jatkinson1000 and @kthyng,
Thank you for handling our submission. We have made the following updates to our repository:
- Reduced the paper to approximately 1000 word (not sure why @editorialbot failed to check the paper info, but the draft was successfully compiled)
- Updated the installation instructions to include PyPI.
- Added documentation on `ReadTheDocs` and jupyter book.
- Included `pytest` for testing the core functions of `pyTCR`.

We look forward to your review.
