### 2025-04-03 #7849 @jatkinson1000
Hi @levuvietphong as I suspected, there seems to be an issue when processing the markdown into a pdf.

It would appear that the script is picking up a `paper.md` in the `docs/` folder instead of the `joss/paper.md`.

Would it be possible to rename the file in `docs/` to something else, e.g. `references.md` or `bibliography.md` to avoid this?

In case it helps, it is possible to add a workflow to your repository to do check the paper - see for example [here](https://github.com/Cambridge-ICCS/FTorch/blob/main/.github/workflows/JOSS_paper_pdf.yml)

I'll take a look over your other changes shortly now that I've worked out what the issue with paper compilation is!
