### 2025-04-03 #7849 @jatkinson1000
Hi @levuvietphong Thank you for responding to the requests in a comprehensive manner!

Hosting on PyPI makes life much easier for your users and will hopefully aid in adoption of the software.

Seeing some test functions in there is also great. It's hard to retrospectively add testing, it should hopefully make life easier going forwards to update these, and it helps to ensure the reliability of the code for yourself and other users.

Having hosted docs is fantastic.
I would suggest linking to the readthedocs page from the "about" section for the GitHub repository.

I'll start inviting reviewers once the paper compiles.
