### 2025-04-03 1ee90b1
doc: update README file and change docs/paper to avoid confusion for paper compilation
