### 2025-04-03 c7d9430
doc: improve API docstring for I/O and plotting functions
