+++ tcr/iodata.py
@@ -12,37 +12,72 @@ import shapefile
 from bs4 import BeautifulSoup
 from tcr.datadir import DATA_DIR
 
+
 def convert_to_mps(values, conversion_factor=0.514444):
-    """Convert values to meters per second using the given conversion factor."""
+    """
+    Convert the provided values to meters per second (m/s) using the specified conversion factor.
+
+    Parameters:
+    -----------
+    values : array-like
+        The input values to be converted.
+    conversion_factor : float, optional
+        The factor used for conversion to meters per second (default is 0.514444).
+
+    Returns:
+    --------
+    array-like
+        The converted values in meters per second (m/s).
+    """
     return values * conversion_factor
 
 
 def filter_by_year(data, years, start_year, end_year):
-    """Filter data by year range and return filtered data with indices."""
+    """
+    Filter the input data based on a specified year range and return the filtered data along with their indices.
+
+    Parameters:
+    -----------
+    data : array-like
+        The input data to be filtered.
+    years : array-like
+        Array of years corresponding to the input data.
+    start_year : int
+        The starting year of the desired range (inclusive).
+    end_year : int
+        The ending year of the desired range (inclusive).
+
+    Returns:
+    --------
+    tuple
+        A tuple containing the filtered data and the indices of the filtered elements.
+    """
     indices = np.where((years >= start_year) & (years <= end_year))[0]
     return data[indices], indices
 
 
 def load_Matlab_data(directory, filename):
     """
-    Load data in Matlab format.
+    Load data from a Matlab (.mat) file.
 
     Parameters:
     -----------
     directory : str
-        Name of the directory where data is stored.
+        The directory containing the Matlab file.
     filename : str
-        Name of the Matlab file.
+        The name of the Matlab file to be loaded.
 
     Returns:
     --------
     dict
-        Dictionary containing the loaded Matlab data.
+        A dictionary containing the data loaded from the Matlab file.
 
     Raises:
     -------
     FileNotFoundError
-        If the specified file does not exist.
+        If the specified file does not exist in the given directory.
+    RuntimeError
+        If an error occurs while loading the Matlab file.
     """
     try:
         current_dir = os.path.dirname(os.path.abspath(__file__))
@@ -63,43 +98,55 @@ def load_netcdf_track_data(
     data_directory=None, model="E3SM-1-0", basin="NA", expmnt="historical"
 ):
     """
-    Load data in NetCDF format from Tropical cyclone downscaling.
+    Load and extract data from a NetCDF file containing tropical cyclone downscaling information.
 
     Parameters:
     -----------
-    data_directory : str
-        Path to the data directory.
-    model : str
-        Name of the CMIP6 model.
-    basin : str
-        Name of the ocean basin (e.g., NA, WP, EP, SH).
-    expmnt : str
-        Name of the model experiment (e.g., historical, ssp585).
+    data_directory : str, optional
+        Path to the directory containing the NetCDF files.
+        If not provided, defaults to the downscaled data directory.
+    model : str, optional
+        Name of the CMIP6 model (e.g., 'E3SM-1-0'). Default is 'E3SM-1-0'.
+    basin : str, optional
+        Name of the ocean basin (e.g., 'NA' for North Atlantic, 'WP' for Western Pacific).
+        Default is 'NA'.
+    expmnt : str, optional
+        Name of the model experiment (e.g., 'historical', 'ssp585').
+        Default is 'historical'.
 
     Returns:
     --------
     ds : xarray.Dataset
-        The loaded NetCDF dataset.
[…]
