### 2025-04-03 #7849 @levuvietphong
Hi @jatkinson1000,

We added the ReadTheDocs link to the "About" section of the repo. In addition, we fixed the paper compilation issue by updating docs/paper.md to paper/bibliography.md. Thanks.
