### 2025-04-08 13b7ac8
doc: add conda installation to README
