### 2025-04-08 c9b15a4
doc: add conda installation to README
