### 2025-04-08 #7849 @jatkinson1000
Thanks @1JunGu We aim for reviews to be completed in 2-4 weeks, though there is some flexibility if circumstances demand.
Once we have 3 reviewers I will open a new [REVIEW] issue where this will take place.
However, you can start looking at the repository and evaluating against the [checklist](https://joss.readthedocs.io/en/latest/review_checklist.html) and other criteria before this.
