### 2025-04-09 67b8d68
release: v1.2.2
- Updated CHANGELOG for v1.2.2
- Bumped version to 1.2.2 in __init__.py
