+++ tcr/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.2a1"
+__version__ = "1.2.2"
