### 2025-04-11 #7849 @jatkinson1000
Hi there @dorchard, @cheginit, @crlna16!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a Python package for predicting tropical cyclone rainfall and associated statistics based on physical modelling.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and I'm happy to answer any questions.
