### 2025-04-17 #8074 @editorialbot
[REVIEW]: pyTCR: A tropical cyclone rainfall model for python
**Submitting author:** @levuvietphong (<a href="http://orcid.org/0000-0001-5558-1023">Phong V.V. Le</a>)
**Repository:** https://github.com/levuvietphong/pyTCR/
**Branch with paper.md** (empty if default branch): joss
**Version:** 1.2.3
