### 2025-05-01 b1ad390
Update open_dataset in iodata.py with context manger
Change
ds = xr.open_dataset(ncfile)
to
with xr.open_dataset(ncfile) as file_handle:
    ds = file_handle.load()
to avoid the lock.
