### 2025-05-01 #8074 @fzhao70
# Review of pyTCR
### Summary
pyTCR is a well-designed, user-friendly tool for extracting rainfall and wind data along tropical cyclone tracks. Its scientific foundation is solid, and installation and usage are straightforward. I ran the provided example and confirmed that the outputs are correct. The accompanying manuscript satisfies the JOSS publication criteria.

### Recommendations
1. Prevent file locking
When opening multiple tracks_*.nc files, the files remain locked and can trigger “permission denied” errors on subsequent accesses. I recommend updating iodata.py to use a context manager:

From :
```
ds = xr.open_dataset(ncfile)
```
To :
```
with xr.open_dataset(ncfile) as file_handle:
     ds = file_handle.load()
```

This change ensures files are closed immediately after loading. I have submitted a pull request with this fix.

2. Make data directory configurable
Allow users to specify the path to their data directory (e.g., via a function argument or configuration setting) rather than hard-coding it. This will improve flexibility for different workflows and data layouts.

3. Document track file structure
Add a brief section in the documentation that describes the schema and contents of the track files (tracks_*.nc). A clear overview of variables, dimensions, and metadata will help new users understand and adapt the data for their purposes.

### Conclusion
Overall, pyTCR is a valuable and robust open-source toolkit. It is good for publication in JOSS.
