### 2025-05-01 #8074 @jatkinson1000
Thanks so much for your review @fzhao70

We [encourage reviewers to open suggestions as issues on the repository](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#guiding-principles) if possible and link to this review issue so we can keep track of the progress. Would you be able to do this?

@levuvietphong can you take a look at making the changes suggested above?
