### 2025-05-04 98b8167
Update iodata.py
Remove leftover commands (xarray load) to clean up the code
