+++ tcr/iodata.py
@@ -156,7 +156,6 @@ def load_netcdf_track_data(
         ncfile = glob.glob(ncfilename)[0]
         with xr.open_dataset(ncfile) as file_handle:
             ds = file_handle.load()
-        #ds = xr.open_dataset(ncfile)
         lat_trks = np.nan_to_num(ds['lat_trks'].values)
         lon_trks = np.nan_to_num(ds['lon_trks'].values)
         n_trk = np.nan_to_num(ds['n_trk'].values)
@@ -312,7 +311,6 @@ def load_tracks_GCMs(
     ncfilename = os.path.join(data_directory, expmnt, f"tracks_{basin}_{model}_*.nc")
     try:
         ncfile = glob.glob(ncfilename)[0]
-        #ds = xr.open_dataset(ncfile)
         with xr.open_dataset(ncfile) as file_handle:
             ds = file_handle.load()
         lon_trks = ds['lon_trks'].values
