### 2025-05-06 levuvietphong/pyTCR#18 @1JunGu
The incomplete function of Parallel method.
I noticed the function:
https://github.com/levuvietphong/pyTCR/blob/53c5e0fa781c7e09aa438aea419a6a5cf3a844c8/tcr/parallel.py#L10-L20

appears to be incomplete (e.g., as it lacks the necessary import for `mpi4py`). Additionally, this functions seemd not to be used anywhere in this version.

Considering this function as a method to accelerate data processing, delete this or compelete its implementation in a future version is ok. This change wouldn't affect the core scientific functionality of this software.

openjournals/joss-reviews#8074
