### 2025-05-06 levuvietphong/pyTCR#19 @1JunGu
Similar I/O methods for loading track data
During my review and replicating the examples in Juypter Notebooks, I noticed that there are two similar methods to load track data from NetCDF files in `iodata.py`:
1. https://github.com/levuvietphong/pyTCR/blob/53c5e0fa781c7e09aa438aea419a6a5cf3a844c8/tcr/iodata.py#L97
2. https://github.com/levuvietphong/pyTCR/blob/53c5e0fa781c7e09aa438aea419a6a5cf3a844c8/tcr/iodata.py#L278

There are only some differences:
1. **Return Values**: The latter method (`load_track_from_netcdf`) returns significantly more variables compared to the former (`load_netcdf_track`).
2. **NaN Handling**: The former method uses `np.nan_to_num` to replace `NaN` values with a default value, presumably to avoid noise in subsequent analysis, whereas the latter does not appear to handle `NaN` values explicitly.

Could the authors clarify the rationale behind having two separate methods with overlapping functionality? Is there a specific use case or context for each method that justifies maintaining both? Additionally, it might be helpful to document these differences in the code or README to guide users on which method to use.

openjournals/joss-reviews#8074
