### 2025-05-06 73925c4
Create constants.py
1. Create a new python file to save physical constants.
