+++ tcr/constants.py
@@ -1,0 +1,2 @@
+#physical constants used in this repo.
+import os, sys
