### 2025-05-06 levuvietphong/pyTCR#20 @1JunGu
Wrong Earth angular speed calculation
In this software, the Earth angular speed $\Omega$ is a critical physical constants used to calculate the Coriolis Parameter.
However, the $\Omega$ appears to be calculated incorrectly here:
https://github.com/levuvietphong/pyTCR/blob/53c5e0fa781c7e09aa438aea419a6a5cf3a844c8/tcr/wind.py#L773

And also subsequent Coriolis Parameter:
https://github.com/levuvietphong/pyTCR/blob/53c5e0fa781c7e09aa438aea419a6a5cf3a844c8/tcr/wind.py#L173

Therefore, I recommend correcting this error and reevaluating the results in provided examples to ensure accuracy.

openjournals/joss-reviews#8074
