### 2025-05-06 levuvietphong/pyTCR#20 @1JunGu
Also, I noticed that the calculation of the Coriolis Parameter uses an assupmtion that replaces $sin(\phi)$ with $|\phi|$ at:
https://github.com/levuvietphong/pyTCR/blob/53c5e0fa781c7e09aa438aea419a6a5cf3a844c8/tcr/wind.py#L1473

This approximation may not be appropriate, even if $\phi$ is close to 0 (near the equator). Moreover, TC activity predominantly occurs in the subtropics, where this assumption could introduce errors and contrasts with other calculations in other locations

I recommend revisiting this approximation and using $sin(\phi)$ directly to ensure accuracy in the calculation.
