### 2025-05-06 2983c19
1. Replace the inconsistent or chaotic usage of constants throughout the codebase with a uniform approach.   2. Correct the calculation of OMEGA.
