### 2025-05-07 #8074 @1JunGu
# Review report for pyTCR

## Summary
Following the JOSS review guidelines and checklists, I find pyTCR to be a user-friendly and well-structured Python software package. It addresses the challenge of estimating tropical cyclone (TC)-induced rainfall by utilizing vertical wind data within the TC vortex. In contrast to numerical weather prediction models, which require expensive computing resources to solve primitive equations and implement physical parameterizations, pyTCR provides a simpler yet effective method for rainfall estimation. This approach shows considerable potential, particularly when coupled with risk models, for delivering prompt and quantified rainfall assessments.

After thorough evaluation of the code, documentation, examples, and the software paper, I recommend accepting this submission once the following issues have been addressed:

- [ ] 1. Resolve issue: levuvietphong/pyTCR#18
- [ ] 2. Resolve issue: levuvietphong/pyTCR#19
- [ ] 3. Resolve issue: levuvietphong/pyTCR#20

Additionally, I have submitted a pull request (levuvietphong/pyTCR#21) that introduces a centralized file for physical constants to enhance code readability, maintainability, and consistency.

## Other Comments on the Software Paper
1.  **Line 31:** The full name "Tropical Cyclones" and its abbreviation "(TCs)" are repeated; this redundancy should be revised.
2.  **Line 74:** The word "Figure" is incorrectly capitalized following a comma.
