### 2025-05-07 05db790
feat: add more physical constants and model parameters to tcr/constants.py for consistence across all functions
