+++ tcr/wind.py
@@ -10,7 +10,7 @@ import glob
 import numpy as np
 from tcr import terrain_boundary as tcr_tb
 from tcr import iodata as tcr_io
-from tcr import constants as tcr_cons
+from tcr import constants as tcr_const
 from tcr.datadir import DATA_DIR
 
 
@@ -708,7 +708,8 @@ def get_translation_speeds_full(lats, longs, vs=None, u850=None, v850=None):
 def calculate_upward_velocity_field(
     latitude, longitude, velocity, radius_storm, velocity_secondary,
     radius_storm_secondary, ut, vt, us, vs, plat, plong, h, hx, hy,
-    deltar=2, timeresw=2, Htrop=4000, radcity=300, wprofile=3
+    deltar=tcr_const.deltar, timeresw=tcr_const.timeresw, Htrop=tcr_const.Htrop,
+    radcity=tcr_const.radcity, wprofile=tcr_const.wprofile
 ):
     """
     Calculate the spatial distribution of vertical velocity.
@@ -762,17 +763,14 @@ def calculate_upward_velocity_field(
     sy = 1 if ngrid == 1 else np.max(np.shape(plat))
 
     plong = plong + 360 if plong[0] < 0 else plong
-
-    pifac = tcr_cons.RAD2DEG
-    dfac = 60 * 1.852
-
-    # scale factor converting nautical miles to meters (using 0.25 degree resolution).
-    sfac = 1 / (0.25 * 60 * 1852)
-    knotfac = tcr_cons.KNOTS2MPS  # convert knots to m/s (1 knots = 0.5144 m/s)
+    res = 0.25  # resolution in degrees
+    pifac = tcr_const.RAD2DEG       # convert radians to degrees (pi/180)
+    dfac = tcr_const.DEG2KM         # factor converting degree to km
+    sfac = dfac / res * 1000        # resolution in degree to meter
+    knotfac = tcr_const.KNOTS2MPS   # convert knots to m/s (1 knots = 0.5144 m/s)
     ut = ut * knotfac   # convert to m/s
     vt = vt * knotfac   # convert to m/s
-    #omega = math.pi / (6 * 3600)
-    omega = tcr_cons.OMEGA  # Earth angular velocity parameter (rad/s)
+    omega = tcr_const.OMEGA  # Earth angular velocity parameter (rad/s)
     latfac = (latitude[0] / (abs(latitude[0]) + 1e-8) if latitude.ndim == 1
               else latitude[0, 0] / (abs(latitude[0, 0]) + 1e-8))
 
@@ -853,8 +851,9 @@ def calculate_upward_velocity_field(
 
 def calculate_wind_time_series(latitude, longitude, velocity, radius_storm,
                                velocity_secondary, radius_storm_secondary, ut, vt,
-                               plat, plong, timeres, timelength=96, wheight=30,
-                               wprofile=3, radcity=300):
+                               plat, plong, timeres, timelength=tcr_const.timelength,
+                               wheight=tcr_const.wheight, wprofile=tcr_const.wprofile,
+                               radcity=tcr_const.radcity):
     """
     Calculate the time series of wind speed and direction at specified points
     of interest.
@@ -1058,8 +1057,9 @@ def calculate_wind_time_series(latitude, longitude, velocity, radius_storm,
 def calculate_upward_velocity_time_series(
     latitude, longitude, velocity, radius_storm,
     velocity_secondary, radius_storm_secondary, ut, vt, us, vs, plong, plat,
[…]
