### 2025-05-07 f08f077
feat: remove unused parallel.py file
The domain_decompose function was not being used anywhere in the current version of the codebase, so it has been removed to reduce clutter and improve maintainability.

This resolves #18.
