### 2025-05-07 7b24122
fix: correct the scale factor converting degree to SI unit and update notebooks
