### 2025-05-07 levuvietphong/pyTCR#19 @levuvietphong
We've merged these two functions into a single one to improve consistency across the codebase, reduce redundancy, and simplify maintenance moving forward. Thank you.
