### 2025-05-07 50370ab
feat: combine two functions loading GCM data into a single function for consistency across the codebase. Notebooks are also updated.
This resolves #19.
