### 2025-05-08 330ccb9
fix: update load_tracks_GCMs in ex3 for loading netcdf track file
