+++ tcr/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.2"
+__version__ = "1.2.3a1"  # Version of the TCR package
