### 2025-05-08 levuvietphong/pyTCR#20 @levuvietphong
This has been resolved in commit 2983c1983df5cf2e0e31689d3e88c65f4497ae13
