### 2025-05-08 4edd458
docs: update API reference for jupyter-book build
