### 2025-05-08 50e0b65
feat: add a config option that allows users to set data directory path instead of using the default. This resolves issue #16.     - Users can set/change the data directory path in the `config.yaml`` file     - tcr/constants.py was deleted and replaced with tcr/parameters.py     - Notebook examples were updated to reflect the changes made in the codebase
