+++ tcr/rainfall.py
@@ -8,16 +8,16 @@ import numpy as np
 from tcr import wind as tcr_wind
 from tcr import terrain_boundary as tcr_tb
 from tcr import iodata as tcr_io
-from tcr import constants as tcr_const
-from tcr.datadir import DATA_DIR
+from tcr import parameters as tcr_params
+from tcr.datadir import BASE_DATA_DIR, DOWNSCALED_DATA_DIR
 
 
 def calculate_rainfall_rate(
     nt, latitude, longitude, radius_storm, velocity, radius_storm_secondary,
     velocity_secondary, ut, vt, u850, v850, months, days, hours, monthplot,
-    dayplot, hourplot, extent=None, shapefile=None, magfac=tcr_const.magfac,
-    deltax=tcr_const.deltax, deltay=tcr_const.deltay, dellatlong=tcr_const.dellatlong,
-    q900=tcr_const.q900, eprecip=tcr_const.eprecip, wrad=tcr_const.wrad
+    dayplot, hourplot, extent=None, shapefile=None, magfac=tcr_params.magfac,
+    deltax=tcr_params.deltax, deltay=tcr_params.deltay, dellatlong=tcr_params.dellatlong,
+    q900=tcr_params.q900, eprecip=tcr_params.eprecip, wrad=tcr_params.wrad
 ):
     """
     Computes the surface rain rate distribution (mm/hr) for a specified storm
@@ -100,9 +100,9 @@ def calculate_rainfall_rate(
             bxmin, bxmax, bymin, bymax = tcr_io.get_bbox_from_shapefile(
                 shapefile)
 
-    rhoa_over_rhol = tcr_const.RHOA_OVER_RHOL  # ratio of air density over water density
-    pifac = tcr_const.RAD2DEG  # convert radians to degrees (pi/180)
-    knotfac = tcr_const.KNOTS2MPS  # convert knots to m/s (1 knots = 0.5144 m/s)    
+    rhoa_over_rhol = tcr_params.RHOA_OVER_RHOL  # ratio of air density over water density
+    pifac = tcr_params.RAD2DEG  # convert radians to degrees (pi/180)
+    knotfac = tcr_params.KNOTS2MPS  # convert knots to m/s (1 knots = 0.5144 m/s)    
     nrm, mrm = np.shape(radius_storm)
     rfac = magfac * (1+np.zeros((nrm, mrm)))
 
@@ -191,9 +191,9 @@ def calculate_rainfall_rate(
 def calculate_etr_swath(
     nt, latitude, longitude, radius_storm, velocity, radius_storm_secondary,
     velocity_secondary, ut, vt, u850, v850, extent=None, shapefile=None,
-    magfac=1, deltax=tcr_const.deltax, deltay=tcr_const.deltay,
-    dellatlongs=tcr_const.dellatlongs, q900=tcr_const.q900, timeres=tcr_const.timeres,
-    wrad=tcr_const.wrad, eprecip=tcr_const.eprecip
+    magfac=1, deltax=tcr_params.deltax, deltay=tcr_params.deltay,
+    dellatlongs=tcr_params.dellatlongs, q900=tcr_params.q900, timeres=tcr_params.timeres,
+    wrad=tcr_params.wrad, eprecip=tcr_params.eprecip
 ):
     """
     Calculate the distribution of event total rainfall for a given individual
@@ -261,7 +261,7 @@ def calculate_etr_swath(
                 shapefile)
 
     # Constants
-    rhoa_over_rhol = tcr_const.RHOA_OVER_RHOL  # rho_air / rho_liquid
+    rhoa_over_rhol = tcr_params.RHOA_OVER_RHOL  # rho_air / rho_liquid
 
     # Initialize variables
[…]
