### 2025-05-08 2cdda5d
docs: update version to 1.2.3b1
