+++ tcr/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.2.3a2"  # Version of the TCR package
+__version__ = "1.2.3b1"  # Version of the TCR package
