### 2025-05-08 59f26e8
fix: unindent bug in datadir docstring due to bullet points
