### 2025-05-19 #8074 @dorchard
# Summary

pyTCR provides a Python software package for estimating the rainfall driven by tropical cyclones, as well as analysis and interpreting the data. The model is driven by outputs from either observations, reanalysis data, or GCM simulations of tropical cyclone tracks.

The package is well described and provides a comprehensive set of notebooks to systematically work through.  This seems like a very useful package and will be of benefit to the community and I think it should be published in JOSS.

# Main review points

I have two main points that I have left unchecked in the checklist which I think could be addressed before acceptance.

Firstly, is that the documentation of the API for this package not gathered into a single place. There is a readthedocs site but it focuses on the background science and installation, then points to notebooks for usage. Some documentation on the API would be useful; there is good documentation within the code itself (with docstrings for all methods I looked at), which is great. Even better would then be to auto generate browseable documentation from this.

Secondly, there are some claims of improved efficiency in the JOSS paper but it should be made more clear with what this is being compared.

- l.11 "increases computational efficiency" I believe is referring to in comparison with the "several legacy physical models" earlier in the sentence, but this was a bit vague. There doesn't appear to be any evidence of this (and the models were not explicitly stated).

- l.23-24 also make a claim about efficiency but there doesn't appear to be any evidence given.

I think generally the point is that the approached here is computational much more simple than running a whole GCM (see l.49) and I buy that argument without data. Perhaps there just needs a bit of tweaking to the wording above so things are a bit more clear, with a bit of explanation as to why the efficiency is greatly improved here.

# Other comments

- It would be useful to state in the Installation section the minimum version of Python that can be used here and/or which versions it is known to work with. I had problems with 3.9 but 3.11 worked fine.

- In the bibliography, most of the paper titles have been forced to lower case, e.g., "Tropical cyclone activity in the high-resolution community earth system model and the impact of ocean coupling." should at least have "Community Earth System Model" in it. You can do this in the .bib file by encasing each upper case letter in { } or in many cases you can just do this for the whole title, e.g., `title = {{Tropical cyclone activity...}}`
