### 2025-05-19 #8074 @jatkinson1000
Thanks @dorchard

API Docs:
I believe there is an API reference on the RTD page here: https://pytcr.readthedocs.io/en/latest/api/pyTCR.html
as this is something I requested pre-review. Or are you looking for something more/improvements on this?

Performance Claims:
Yes - we ask authors to cite alternative approaches/pieces of software in their paper.
Performance claims should be justified according to the JOSS criteria. In this case there are no hard claims, but reference should be made to the alternative approaches and exactly what is meant by "increased efficiency" - if indeed it is compared to running a full GCM or regional model then adjusting the wording to clarify that would be sufficient.

Python version:
Setting software requirements to a minimum working python version would be useful - please open an issue for this @dorchard

Typesetting:
This should be caught in the finalisation stages, but thank you for raising it here. It should be amended to capitalise appropriately.

@levuvietphong are you able to look at the points @dorchard has raised?
