### 2025-05-19 levuvietphong/pyTCR#23 @dorchard
Specify minimum Python 3 version / tested Python versions
Great work here!

It would be useful to state in the Installation section the minimum version of Python that can be used here and/or which versions it is known to work with. I had problems with 3.9 but 3.11 worked fine.
