### 2025-05-19 #8074 @dorchard
* Re API docs. My apologies. I somehow missed this. I'm happy with what is there (have updated my checklist).
* Performance claims. What you suggests sounds good re clarifying what is being compared in the performance claims.
* Have made an issue for the Python version info here: https://github.com/levuvietphong/pyTCR/issues/23
