### 2025-05-19 50a60d4
docs: update bibliography formatting to prevent forced lowercase letters
