+++ joss/paper.bib
@@ -7,7 +7,7 @@
 
 @article{Lin:2023,
     author = {Lin, Jonathan and Rousseau-Rizzi, Raphael and Lee, Chia-Ying and Sobel, Adam},
-    title = {An Open-Source, Physics-Based, Tropical Cyclone Downscaling Model With Intensity-Dependent Steering},
+    title = {{An Open-Source, Physics-Based, Tropical Cyclone Downscaling Model With Intensity-Dependent Steering}},
     journal = {Journal of Advances in Modeling Earth Systems},
     volume = {15},
     number = {11},
@@ -22,7 +22,7 @@
 
 @article {Lu:2018,
     author = "Ping Lu and Ning Lin and Kerry Emanuel and Daniel Chavas and James Smith",
-    title = "Assessing Hurricane Rainfall Mechanisms Using a Physics-Based Model: Hurricanes Isabel (2003) and Irene (2011)",
+    title = "{Assessing Hurricane Rainfall Mechanisms Using a Physics-Based Model: Hurricanes Isabel (2003) and Irene (2011)}",
     journal = "Journal of the Atmospheric Sciences",
     year = "2018",
     publisher = "American Meteorological Society",
@@ -36,7 +36,7 @@
 
 @article{Zhu:2013,
     author = {Zhu, Laiyin and Quiring, Steven M. and Emanuel, Kerry A.},
-    title = {Estimating tropical cyclone precipitation risk in Texas},
+    title = {{Estimating tropical cyclone precipitation risk in Texas}},
     journal = {Geophysical Research Letters},
     volume = {40},
     number = {23},
@@ -50,8 +50,7 @@
 
 @Article{Eyring:2016,
     AUTHOR = {Eyring, V. and Bony, S. and Meehl, G. A. and Senior, C. A. and Stevens, B. and Stouffer, R. J. and Taylor, K. E.},
-    TITLE = {Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6)
-    experimental design and organization},
+    TITLE = {{Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6) experimental design and organization}},
     JOURNAL = {Geoscientific Model Development},
     VOLUME = {9},
     YEAR = {2016},
@@ -63,7 +62,7 @@
 
 @article{Lopez:2024,
     author = {Hosmay Lopez  and Sang-Ki Lee  and Robert West  and Dongmin Kim  and Gregory R. Foltz  and Ghassan J. Alaka  and Hiroyuki Murakami },
-    title = {Projected increase in the frequency of extremely active Atlantic hurricane seasons},
+    title = {{Projected increase in the frequency of extremely active Atlantic hurricane seasons}},
     journal = {Science Advances},
     volume = {10},
     number = {46},
@@ -76,7 +75,7 @@
 
 @Article{Reed:2022,
     author={Reed, Kevin A. and Wehner, Michael F. and Zarzycki, Colin M.},
-    title={Attribution of 2020 hurricane season extreme rainfall to human-induced climate change},
+    title={{Attribution of 2020 hurricane season extreme rainfall to human-induced climate change}},
     journal={Nature Communications},
     year={2022},
     month={Apr},
@@ -91,7 +90,7 @@
 
 @Article{Krichene:2023,
     author={Krichene, Hazem and Vogt, Thomas and Piontek, Franziska and Geiger, Tobias and Sch{\"o}tz, Christof and Otto, Christian},
-    title={The social costs of tropical cyclones},
+    title={{The social costs of tropical cyclones}},
     journal={Nature Communications},
     year={2023},
     month={Nov},
@@ -107,7 +106,7 @@
 @Article{Young:2024,
     author={Young, Rachel
     and Hsiang, Solomon},
-    title={Mortality caused by tropical cyclones in the United States},
[…]
