### 2025-05-19 418d1c8
docs: update manuscript to address the concerns on efficiency raised by reviewers
