### 2025-05-19 35e98be
docs: update README.md to add minimum required version for python. This resolves and closes #23.
