+++ README.md
@@ -40,6 +40,10 @@ pyTCR is a physics-based model developed in python to estimate rainfall induced 
     ```sh
     conda install -c conda-forge pyTCR
     ```
+
+    > 💡 **Note:**
+    > *Python 3.11 or higher is required for installing pyTCR.*
+    
 2. To get the latest (unstable) development version from GitHub:
     ```sh
     git clone https://github.com/levuvietphong/pyTCR.git
@@ -48,10 +52,10 @@ pyTCR is a physics-based model developed in python to estimate rainfall induced 
     ```
 
     > 💡 **Note:**
-    > *This installs the most recent updates, which may include new features and bug fixes that are not yet available in the stable release. Cloning the repository also gives you access to the tutorial in the notebooks folder.*
+    > *This installs the most recent updates, which may include new features and bug fixes that are not yet available in the stable release.*
 
 ## :arrow_forward: Getting Started
-This repository provides a collection of [jupyter notebooks](https://github.com/levuvietphong/pyTCR/tree/main/notebooks) designed to facilitate the use of PyTCR for generating and visualizing rainfall and wind speed induced by TCs. These notebooks cover various use cases and data sources, providing comprehensive guidance for users.
+This repository provides a collection of [jupyter notebooks](https://github.com/levuvietphong/pyTCR/tree/main/notebooks) designed to facilitate the use of PyTCR for generating and visualizing rainfall and wind speed induced by TCs. To access the tutorials, clone the repository locally. These notebooks cover various use cases and data sources, providing comprehensive guidance for users. 
 
 0. [Downloading Tropical Cyclone Data:](./notebooks/ex0_download_tracks_from_cmip6.ipynb) illustrates the process of downloading CMIP6 tropical cyclone tracks that have undergone downscaling using the tropical cyclone downscaling [model](https://github.com/linjonathan/tropical_cyclone_risk).
 
