### 2025-05-19 #8074 @levuvietphong
@dorchard @jatkinson1000: Thank you for your feedback. We've updated the code and manuscript to address your comments:

1. **Performance claims**: We agree with this point - there was no direct performance comparison between pyTCR and GCMs. To avoid any confusion, we have removed the term "efficiency increase" and revised the `Summary` section of the manuscript as follows:

    > Line11: It integrates the capabilities of several legacy physical models into a reduced-complexity framework, enabling rapid estimation of tropical cyclone (TC) rainfall consistent with the large-scale environment.
    > Line 23-24: As a result, `pyTCR` provides a computationally efficient approach for capturing extreme TC rainfall events at the tail of the distributions from limited datasets.

2. **Python version**: We've updated the README.md to specify minimum required Python version in the INSTALLATION section. See commit [35e98be](https://github.com/levuvietphong/pyTCR/commit/35e98be18eb8984771fafbd13b00f9d46a3588d1).

3. **Bibliography**: We've updated the .bib file to  avoid automatic lowercasing.
