### 2025-05-19 #8074 @jatkinson1000
@levuvietphong I suggest also adding
```toml
requires-python = ">=3.11"
```
to your `pyproject.toml` under the `[project]` specification.

JOSS guidance asks that the paper contain [A list of key references, including to other software addressing related needs](https://joss.readthedocs.io/en/latest/paper.html#what-should-my-paper-contain). Please could you add in references to the "legacy models" that you are providing an alternative for.
