### 2025-05-19 229c7c4
docs: update references for legacy physical models
