+++ joss/paper.bib
@@ -382,3 +382,17 @@
     year = {2018},
     month = {apr}
 }
+
+@article{Langousis:2009,
+    author = {Langousis, Andreas and Veneziano, Daniele},
+    title = {Theoretical model of rainfall in tropical cyclones for the assessment of long-term risk},
+    journal = {Journal of Geophysical Research: Atmospheres},
+    volume = {114},
+    number = {D2},
+    pages = {},
+    keywords = {precipitation, tropical cyclones, rainfall},
+    doi = {https://doi.org/10.1029/2008JD010080},
+    url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2008JD010080},
+    year = {2009}
+}
+
