### 2025-05-19 9b6184c
Update CI.yml to use Python version 3.11 reflecting the minimum supported version from the documentation
