### 2025-05-19 #8074 @levuvietphong
Thanks @jatkinson1000 for the pull request. It has been merged into  the `main` branch.
References have been updated for the manuscript as well.
