### 2025-05-21 ea928a5
docs: update to 1.2.3b2 and minimum version requirement to readthedocs
