### 2025-05-21 1ce01f1
docs: update documentation for v1.2.3b2
