+++ docs/datatypes.md
@@ -1,13 +1,10 @@
 # Data
 
-`pyTCR` processes tropical cyclone (TC) input data obtained from observational sources or downscaled outputs generated by the TC downscaling model {cite}`Lin:2023`. 
-The TC downscaling model is enforced by large-scale environmental conditions obtained from Earth System Models (ESMs). 
-`pyTCR` supports the `NetCDF` format for data from these two sources. 
-Below, we outline the key variables and dimensions relevant to each source.
+`pyTCR` processes tropical cyclone (TC) input data obtained from observational sources or downscaled outputs generated by the TC downscaling (TCD) model {cite}`Lin:2023`. The TCD model is enforced by large-scale environmental conditions obtained from Earth System Models (ESMs). `pyTCR` supports the `NetCDF` format for data from these sources. We outline the key variables and dimensions relevant to each source below.
 
-## 1. TC downscaling model outputs
+## 1. Tropical cyclone downscaling model outputs
 
-The outputs of the TC downscaling model contains TC track data. The dataset covers the historical period from 1964 to 2014 and includes information on TC position, intensity, and dynamic variables. Moreover, it provides metadata that describes the temporal and spatial dimensions of the data.
+The outputs of the TCD model contains TC track data. The dataset covers the historical period from 1964 to 2014 and includes information on TC position, intensity, and dynamic variables. Moreover, it provides metadata that describes the temporal and spatial dimensions of the data.
 
 ### Dimensions
 
@@ -18,7 +15,6 @@ The outputs of the TC downscaling model contains TC track data. The dataset cove
 | `year`  | 51   | Simulation years (1964–2014)                     |
 | `month` | 12   | Calendar months (January–December)               |
 | `basin` | 7    | Ocean basin (encoded as strings)                 |
-
 
 ### Key Variables
 
@@ -51,9 +47,8 @@ The outputs of the TC downscaling model contains TC track data. The dataset cove
 | ----------------- | -------------------- | ----------------------------- | ----- |
 | `seeds_per_month` | (year, basin, month) | Number of random seeds placed | -     |
 
-
 ## 2. Observations - International Best Track Archive for Climate Stewardship (IBTrACS)
 
 The IBTrACS project is the most complete global observations of tropical cyclones dataset. It merges recent and historical tropical cyclone data from multiple agencies to create a unified, publicly available, best-track dataset that improves inter-agency comparisons.
 
-See the [Technical Documentation of IBTrACS](https://www.ncei.noaa.gov/products/international-best-track-archive) for more information
+See the [Technical Documentation of IBTrACS](https://www.ncei.noaa.gov/products/international-best-track-archive) for more information.
