+++ joss/paper.bib
@@ -204,6 +204,32 @@
     journal = {Environmental Research Letters},
 }
 
+@article { Emanuel:2006a,
+    author = "Kerry Emanuel",
+    title = "Climate and Tropical Cyclone Activity: A New Model Downscaling Approach",
+    journal = "Journal of Climate",
+    year = "2006",
+    publisher = "American Meteorological Society",
+    address = "Boston MA, USA",
+    volume = "19",
+    number = "19",
+    doi = "10.1175/JCLI3908.1",
+    pages=      "4797 - 4802",
+    url = "https://journals.ametsoc.org/view/journals/clim/19/19/jcli3908.1.xml"
+}
+@article {Emanuel:2006b,
+    author = "Kerry Emanuel and Sai Ravela and Emmanuel Vivant and Camille Risi",
+    title = "A Statistical Deterministic Approach to Hurricane Risk Assessment",
+    journal = "Bulletin of the American Meteorological Society",
+    year = "2006",
+    publisher = "American Meteorological Society",
+    address = "Boston MA, USA",
+    volume = "87",
+    number = "3",
+    doi = "10.1175/BAMS-87-3-299",
+    pages=      "299 - 314",
+    url = "https://journals.ametsoc.org/view/journals/bams/87/3/bams-87-3-299.xml"
+}
 @article {Emanuel:2008,
     author = "Kerry Emanuel and Ragoth Sundararajan and John Williams",
     title = "{Hurricanes and Global Warming: Results from Downscaling IPCC AR4 Simulations}",
