### 2025-05-29 #8074 @levuvietphong
@jatkinson1000 @dorchard: Thank you for your feedback. We appreciate your sensitivity to this issue.

`pyTCR` uses outputs from a synthetic tropical cyclone downscaling (`TCD`) model to generate along-track, TC-driven rainfall events. The `TCD` model combines statistical track generation with simple TC intensity model (a one-dimensional ODE) that can be run a large number of times. In essence, `pyTCR` and the `TCD` models leverage GCM outputs and increase the effective sample size using simple embedded models to produce statistically robust estimates of the probability distributions of storms under different climate conditions from GCMs.

We have revised the `Statement of need` section to clarify this point. Additionally, we replaced the term "efficient" with "fast" or "tractable" in the manuscript, as these better reflect the intended meaning. The second paragraph of this section reads as follows:

> The ability of ESMs to simulate climate extremes has substantially improved over the past few decades. These models have become key tools used for exploring the effect of global warming on precipitation and climate variability (Emanuel 2021; Le et al. 2021, 2023). While high-resolution ESMs have enhanced the representation of TCs (Haarsma et al. 2016; Li and Sriver 2018), they remain computationally intensive to conduct, so that only a limited number of simulations can be performed. This constrains their application in TC rainfall risk analysis, which requires extensive sampling of extreme events (Emanuel, Sundararajan, and Williams 2008). pyTCR addresses this need with an easy-to-use and fast tool that facilitates TC-driven rainfall analysis across scales. Specifically, it leverages a synthetic downscaling approach that combines statistical track generation with simple deterministic intensity modeling. This approach uses thermodynamic and kinematic statistics derived from ESM outputs or reanalysis data to generate large (~103-104) numbers of synthetic TCs (Emanuel et al. 2006; Lin et al. 2023). As a result, pyTCR produces statistically robust estimates of the probability distributions of storms for risk assessment.

The draft PDF file has been updated [here](https://github.com/levuvietphong/pyTCR/actions/runs/15333919596/artifacts/3224307036).

Thanks.
