### 2025-05-30 #8074 @jatkinson1000
Thanks for your patience and updates @levuvietphong.
My reading is that PyTCR allows users to use the results from a small number of computationally expensive ESM or reanalysis runs to generate a range of statistical probabilities of rainfall by using the Lin model.
You are taking what might be a single number (or small sample) from an ESM run(s), and adding statistical 'error bars'.
Is that the correct interpretation?
If so I think this is clear.

@dorchard are you happy with the changes in commits: https://github.com/levuvietphong/pyTCR/commit/b0f97203bb4aa1138695c45af5dd34e7d32f6732 and https://github.com/levuvietphong/pyTCR/commit/57e472c1038587c9ace5b0355c8c31b21b8f2a32 ?
