### 2025-05-30 #8074 @levuvietphong
> Thanks for your patience and updates [@levuvietphong](https://github.com/levuvietphong). My reading is that PyTCR allows users to use the results from a small number of computationally expensive ESM or reanalysis runs to generate a range of statistical probabilities of rainfall by using the Lin model. You are taking what might be a single number (or small sample) from an ESM run(s), and adding statistical 'error bars'. Is that the correct interpretation? If so I think this is clear.

@jatkinson1000 Yes, your interpretation is correct. To elaborate, the Lin model first reads large-scale inputs from GCMs (e.g., scenarios or ensemble members) or reanalysis data. Then seeds are randomly placed in space and time for genesis. Next, these seeds evolve in space and time according to a track model that is driven by large-scale environmental condition. Finally, an intensity model is applied to simulate storm intensification.

By increasing the number of seeds, we can generate as many tropical cyclone tracks as needed for sampling of extreme events. These tracks remain consistent with the underlying large-scale environmental conditions provided by the GCMs.

`pyTCR` leverages this approach to produce statistically robust estimates of the probability distributions of storms.
