### 2025-06-09 a0a7b21
Phrasing and typesetting of Mathematical approach section.
