### 2025-06-09 982d9ed
Phrasing and typesetting of Examples section.
