### 2025-06-09 f613657
Fix broken DOI and formatting of bibfile.
