+++ joss/paper.bib
@@ -1,8 +1,8 @@
 @misc{CorralTACCHPC,
-  title = {Corral - {{TACC HPC Documentation}}},
-  urldate = {2025-02-03},
-  howpublished = {https://docs.tacc.utexas.edu/hpc/corral/},
-  file = {/Users/wmobley/Zotero/storage/555KLGBI/corral.html}
+    title = {Corral - {{TACC HPC Documentation}}},
+    urldate = {2025-02-03},
+    howpublished = {https://docs.tacc.utexas.edu/hpc/corral/},
+    file = {/Users/wmobley/Zotero/storage/555KLGBI/corral.html}
 }
 
 @article{Lin:2023,
@@ -30,7 +30,7 @@
     volume = "75",
     number = "7",
     doi = "10.1175/JAS-D-17-0264.1",
-    pages=      "2337 - 2358",
+    pages = "2337 - 2358",
     url = "https://journals.ametsoc.org/view/journals/atsc/75/7/jas-d-17-0264.1.xml"
 }
 
@@ -49,15 +49,15 @@
 }
 
 @Article{Eyring:2016,
-    AUTHOR = {Eyring, V. and Bony, S. and Meehl, G. A. and Senior, C. A. and Stevens, B. and Stouffer, R. J. and Taylor, K. E.},
-    TITLE = {{Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6) experimental design and organization}},
-    JOURNAL = {Geoscientific Model Development},
-    VOLUME = {9},
-    YEAR = {2016},
-    NUMBER = {5},
-    PAGES = {1937--1958},
-    URL = {https://gmd.copernicus.org/articles/9/1937/2016/},
-    DOI = {10.5194/gmd-9-1937-2016}
+    author = {Eyring, V. and Bony, S. and Meehl, G. A. and Senior, C. A. and Stevens, B. and Stouffer, R. J. and Taylor, K. E.},
+    title = {{Overview of the Coupled Model Intercomparison Project Phase 6 (CMIP6) experimental design and organization}},
+    journal = {Geoscientific Model Development},
+    volume = {9},
+    year = {2016},
+    number = {5},
+    pages = {1937--1958},
+    url = {https://gmd.copernicus.org/articles/9/1937/2016/},
+    doi = {10.5194/gmd-9-1937-2016}
 }
 
 @article{Lopez:2024,
@@ -74,79 +74,79 @@
 }
 
 @Article{Reed:2022,
-    author={Reed, Kevin A. and Wehner, Michael F. and Zarzycki, Colin M.},
-    title={{Attribution of 2020 hurricane season extreme rainfall to human-induced climate change}},
-    journal={Nature Communications},
-    year={2022},
-    month={Apr},
-    day={12},
-    volume={13},
-    number={1},
-    pages={1905},
-    issn={2041-1723},
-    doi={10.1038/s41467-022-29379-1},
-    url={https://doi.org/10.1038/s41467-022-29379-1}
+    author = {Reed, Kevin A. and Wehner, Michael F. and Zarzycki, Colin M.},
+    title = {{Attribution of 2020 hurricane season extreme rainfall to human-induced climate change}},
+    journal = {Nature Communications},
+    year = {2022},
+    month = {Apr},
+    day = {12},
+    volume = {13},
+    number = {1},
+    pages = {1905},
+    issn = {2041-1723},
+    doi = {10.1038/s41467-022-29379-1},
+    url = {https://doi.org/10.1038/s41467-022-29379-1}
 }
 
 @Article{Krichene:2023,
-    author={Krichene, Hazem and Vogt, Thomas and Piontek, Franziska and Geiger, Tobias and Sch{\"o}tz, Christof and Otto, Christian},
-    title={{The social costs of tropical cyclones}},
[…]
