+++ joss/paper.bib
@@ -421,3 +421,17 @@
     url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2008JD010080},
     year = {2009}
 }
+
+@article {IBTrACS:2010,
+    author = "Kenneth R. Knapp and Michael C. Kruk and David H. Levinson and Howard J. Diamond and Charles J. Neumann",
+    title = "The International Best Track Archive for Climate Stewardship (IBTrACS): Unifying Tropical Cyclone Data",
+    journal = "Bulletin of the American Meteorological Society",
+    year = "2010",
+    publisher = "American Meteorological Society",
+    address = "Boston MA, USA",
+    volume = "91",
+    number = "3",
+    doi = "10.1175/2009BAMS2755.1",
+    pages=      "363 - 376",
+    url = "https://journals.ametsoc.org/view/journals/bams/91/3/2009bams2755_1.xml"
+}
