### 2025-06-11 0d893d6
docs: remove TACC documentation ref that is not used
