+++ joss/paper.bib
@@ -1,10 +1,3 @@
-@misc{CorralTACCHPC,
-    title = {Corral - {{TACC HPC Documentation}}},
-    urldate = {2025-02-03},
-    howpublished = {https://docs.tacc.utexas.edu/hpc/corral/},
-    file = {/Users/wmobley/Zotero/storage/555KLGBI/corral.html}
-}
-
 @article{Lin:2023,
     author = {Lin, Jonathan and Rousseau-Rizzi, Raphael and Lee, Chia-Ying and Sobel, Adam},
     title = {{An Open-Source, Physics-Based, Tropical Cyclone Downscaling Model With Intensity-Dependent Steering}},
