### 2025-06-11 fe13d80
docs: update CHANGELOG for ver 1.2.3
