### 2025-06-11 3e3d867
docs: update offcial version to 1.2.3
