### 2025-06-14 #8074 @levuvietphong
Hi @jatkinson1000.

Thank you and the reviewers for reviewing our manuscript and software. Below are the version number and DOI details for pyTCR:
- DOI: [10.5281/zenodo.15664074](https://doi.org/10.5281/zenodo.15664074)
- Version number: 1.2.3 (https://github.com/levuvietphong/pyTCR/releases/tag/pyTCR-1.2.3)
