### 2025-06-19 #8074 @kthyng
Congratulations on your new publication @levuvietphong! Many thanks to editor @jatkinson1000 and to reviewers @1JunGu, @dorchard, and @fzhao70 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
